+++ packages/openapi-ts/src/openApi/v3/parser/getModels.ts
@@ -1,27 +1,37 @@
-import type { Model } from '../../common/interfaces/client';
+import type { Client } from '../../../types/client';
 import { reservedWords } from '../../common/parser/reservedWords';
 import { getType } from '../../common/parser/type';
 import type { OpenApi } from '../interfaces/OpenApi';
 import { getModel } from './getModel';
 
-export const getModels = (openApi: OpenApi): Model[] => {
-  if (!openApi.components) {
-    return [];
-  }
+export const getModels = (
+  openApi: OpenApi,
+): Pick<Client, 'models' | 'types'> => {
+  const types: Client['types'] = {};
+  let models: Client['models'] = [];
 
-  let models: Model[] = [];
+  if (!openApi.components) {
+    return {
+      models,
+      types,
+    };
+  }
 
   Object.entries(openApi.components.schemas ?? {}).forEach(
     ([definitionName, definition]) => {
       const definitionType = getType({ type: definitionName });
+      const name = definitionType.base.replace(reservedWords, '_$1');
+      const meta = {
+        $ref: `#/components/schemas/${definitionName}`,
+        name,
+      };
+      types[name] = meta;
       const model = getModel({
         definition,
         isDefinition: true,
-        meta: {
-          $ref: `#/components/schemas/${definitionName}`,
-          name: definitionType.base.replace(reservedWords, '_$1'),
-        },
+        meta,
         openApi,
+        types,
       });
       models = [...models, model];
     },
@@ -35,28 +45,32 @@ export const getModels = (openApi: OpenApi): Model[] => {
       }
 
       const definitionType = getType({ type: definitionName });
+      /**
+       * Prefix parameter names to avoid name conflicts with schemas.
+       * Assuming people are mostly interested in importing schema types
+       * and don't care about this name as much. It should be resolved in
+       * a cleaner way, there just isn't a good deduplication strategy
+       * today. This is a workaround in the meantime, hopefully reducing
+       * the chance of conflicts.
+       *
+       * Example where this would break: schema named `ParameterFoo` and
+       * parameter named `Foo` (this would transform to `ParameterFoo`)
+       *
+       * Note: there's a related code to this workaround in `getType()`
+       * method that needs to be cleaned up when this is addressed.
+       */
+      const name = `Parameter${definitionType.base.replace(reservedWords, '_$1')}`;
+      const meta = {
+        $ref: `#/components/parameters/${definitionName}`,
+        name,
+      };
+      types[name] = meta;
       const model = getModel({
         definition: schema,
         isDefinition: true,
-        meta: {
-          $ref: `#/components/parameters/${definitionName}`,
-          /**
-           * Prefix parameter names to avoid name conflicts with schemas.
-           * Assuming people are mostly interested in importing schema types
-           * and don't care about this name as much. It should be resolved in
-           * a cleaner way, there just isn't a good deduplication strategy
-           * today. This is a workaround in the meantime, hopefully reducing
-           * the chance of conflicts.
-           *
-           * Example where this would break: schema named `ParameterFoo` and
-           * parameter named `Foo` (this would transform to `ParameterFoo`)
-           *
-           * Note: there's a related code to this workaround in `getType()`
-           * method that needs to be cleaned up when this is addressed.
-           */
-          name: `Parameter${definitionType.base.replace(reservedWords, '_$1')}`,
-        },
+        meta,
         openApi,
+        types,
       });
       model.deprecated = definition.deprecated;
       model.description = definition.description || null;
@@ -64,5 +78,8 @@ export const getModels = (openApi: OpenApi): Model[] => {
     },
   );
 
-  return models;
+  return {
+    models,
+    types,
+  };
 };