+++ packages/openapi-ts/src/openApi/v3/parser/getRequiredPropertiesFromComposition.ts
@@ -1,28 +1,60 @@
+import type { Client } from '../../../types/client';
 import type { Model } from '../../common/interfaces/client';
 import { getRef } from '../../common/parser/getRef';
+import { getType } from '../../common/parser/type';
+import type { GetModelFn } from '../interfaces/Model';
 import type { OpenApi } from '../interfaces/OpenApi';
 import type { OpenApiSchema } from '../interfaces/OpenApiSchema';
-import type { getModel } from './getModel';
 
-// Fix for circular dependency
-export type GetModelFn = typeof getModel;
-
-export const getRequiredPropertiesFromComposition = (
-  openApi: OpenApi,
-  required: string[],
-  definitions: OpenApiSchema[],
-  getModel: GetModelFn,
-): Model[] =>
-  definitions
+export const getRequiredPropertiesFromComposition = ({
+  definitions,
+  getModel,
+  openApi,
+  required,
+  types,
+}: {
+  openApi: OpenApi;
+  required: string[];
+  definitions: OpenApiSchema[];
+  getModel: GetModelFn;
+  types: Client['types'];
+}): Model[] => {
+  const requiredProperties = definitions
     .reduce((properties, definition) => {
       if (definition.$ref) {
+        const type = getType({ type: definition.$ref });
+        // avoid circular references if two refs reference each other
+        // if (types[type.base] && types[type.base].$ref === definition.$ref) {
+        //   const schema = getRef<OpenApiSchema>(openApi, definition);
+        //   return [...properties]
+        // }
+
+        const meta = {
+          $ref: definition.$ref,
+          name: type.base,
+        };
+        types[type.base] = meta;
         const schema = getRef<OpenApiSchema>(openApi, definition);
         return [
           ...properties,
-          ...getModel({ definition: schema, openApi }).properties,
+          ...getModel({
+            definition: schema,
+            meta,
+            openApi,
+            types,
+          }).properties,
         ];
       }
-      return [...properties, ...getModel({ definition, openApi }).properties];
+
+      return [
+        ...properties,
+        ...getModel({
+          definition,
+          openApi,
+          parentDefinition: definition,
+          types,
+        }).properties,
+      ];
     }, [] as Model[])
     .filter(
       (property) => !property.isRequired && required.includes(property.name),
@@ -31,3 +63,5 @@ export const getRequiredPropertiesFromComposition = (
       ...property,
       isRequired: true,
     }));
+  return requiredProperties;
+};